% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMPHForecastSuite-package.R
\docType{package}
\name{AMPHForecastSuite-package}
\alias{AMPHForecastSuite-package}
\alias{AMPHForecastSuite}
\title{AMPH Forecast Suite}
\description{
A comprehensive suite of tools and helper functions to build and run
an infectious disease forecasting hub. This package facilitates the
installation and integration of key forecasting packages including
hubverse packages, fable, EpiEstim, EpiNow2, epidatr, and epiforecast.
}
\section{Main Functions}{

\itemize{
  \item \code{\link{install_forecast_packages}}: Install all required forecasting packages
  \item \code{\link{convert_hub_to_fable}}: Convert hub data to fable format
  \item \code{\link{convert_fable_to_hub}}: Convert fable forecasts to hub format
  \item \code{\link{convert_hub_to_epiestim}}: Convert hub data to EpiEstim format
  \item \code{\link{convert_epiestim_to_hub}}: Convert EpiEstim results to hub format
  \item \code{\link{convert_hub_to_epinow2}}: Convert hub data to EpiNow2 format
  \item \code{\link{convert_epinow2_to_hub}}: Convert EpiNow2 results to hub format
  \item \code{\link{get_nhsn_data}}: Retrieve data from NHSN
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ACCIDDA/AMPH_Forecast_Suite}
  \item \url{https://accidda.github.io/AMPH_Forecast_Suite/}
  \item Report bugs at \url{https://github.com/ACCIDDA/AMPH_Forecast_Suite/issues}
}

}
\author{
\strong{Maintainer}: ACCIDDA \email{info@example.com}

}
