% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_hub_repos.R
\name{clone_hub_repos}
\alias{clone_hub_repos}
\title{Clone github forecast repo}
\usage{
clone_hub_repos(disease = "influenza", old_rsv_repo = TRUE, clone_dir = NULL)
}
\arguments{
\item{disease}{Character. One of "influenza", "covid", or "rsv". Default is "influenza".}

\item{old_rsv_repo}{Logical. If TRUE, uses the old RSV forecast hub repository. Default is TRUE. If FALSE, uses the new CDC RSV forecast hub repository.
This parameter is only relevant if disease is "rsv". The RSV forecast hub moved from Hopkins to CDC starting in the 2025-26 season.}

\item{clone_dir}{Character. Directory to clone the repository into. Default is the current working directory.}
}
\value{
Character. The path to the cloned repository.
}
\description{
Clones influenza, covid, or rsv forecast hub repos to use for learning, visualization, etc.
Clones to the current working directory.
}
\examples{
\dontrun{
# Clone the influenza forecast hub repository
flu_repo <- clone_hub_repos(disease = "influenza")
# Clone the COVID-19 forecast hub repository
covid_repo <- clone_hub_repos(disease = "covid")
# Clone the old RSV forecast hub repository
old_rsv_repo <- clone_hub_repos(disease = "rsv", old_rsv_repo = TRUE)
# Clone the new CDC RSV forecast hub repository
new_rsv_repo <- clone_hub_repos(disease = "rsv", old_rsv_repo = FALSE)
}

}
