% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow2_converters.R
\name{convert_epinow2_to_hub}
\alias{convert_epinow2_to_hub}
\title{Convert EpiNow2 Output to Hub Format}
\usage{
convert_epinow2_to_hub(
  epinow2_result,
  location = "unknown",
  target = "inc",
  extract_type = "forecasts"
)
}
\arguments{
\item{epinow2_result}{EpiNow2 result object (output from epinow)}

\item{location}{Character. Location identifier. Default is "unknown".}

\item{target}{Character. Target variable name. Default is "inc".}

\item{extract_type}{Character. Type of data to extract: "forecasts", "estimates", or "both". Default is "forecasts".}
}
\value{
A data frame in hub format
}
\description{
Converts EpiNow2 estimates and forecasts to hub submission format.
}
\examples{
\dontrun{
# Assuming you have EpiNow2 results
# hub_forecasts <- convert_epinow2_to_hub(epinow2_res, location = "US")
}
}
