% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_converters.R
\name{convert_fable_to_hub}
\alias{convert_fable_to_hub}
\title{Convert Fable Forecasts to Hub Format}
\usage{
convert_fable_to_hub(fable_forecast, location = NULL, target = "inc")
}
\arguments{
\item{fable_forecast}{A fable forecast object}

\item{location}{Character. Location identifier. Default is NULL (extract from data).}

\item{target}{Character. Target name for the forecast. Default is "inc".}
}
\value{
A data frame in hub format
}
\description{
Converts fable forecast output to hub submission format.
}
\examples{
\dontrun{
# Assuming you have a fable forecast object
# hub_forecast <- convert_fable_to_hub(fc, location = "US", target = "inc")
}
}
