% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiestim_converters.R
\name{convert_hub_to_epiestim}
\alias{convert_hub_to_epiestim}
\title{Convert Hub Format to EpiEstim Format}
\usage{
convert_hub_to_epiestim(
  hub_data,
  date_col = "date",
  incidence_col = "value",
  location_col = "location",
  location_filter = NULL
)
}
\arguments{
\item{hub_data}{Data frame in hub format with date and incidence columns}

\item{date_col}{Character. Name of the date column. Default is "date".}

\item{incidence_col}{Character. Name of the incidence column. Default is "value".}

\item{location_col}{Character. Name of the location column. Default is "location".}

\item{location_filter}{Character. Location to filter for (if multiple locations). Default is NULL.}
}
\value{
A data frame formatted for EpiEstim with columns: dates, I (incidence)
}
\description{
Converts data from hub format to the format required by EpiEstim package
for estimating time-varying reproduction numbers.
}
\examples{
\dontrun{
hub_data <- data.frame(
  date = seq.Date(as.Date("2023-01-01"), by = "day", length.out = 50),
  location = "US",
  value = rpois(50, 100)
)
epiestim_data <- convert_hub_to_epiestim(hub_data, location_filter = "US")
}
}
