% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_converters.R
\name{convert_hub_to_tsibble}
\alias{convert_hub_to_tsibble}
\title{Convert Hub Format to Fable/tsibble Format}
\usage{
convert_hub_to_tsibble(
  hub_data,
  date_col = "date",
  value_col = "value",
  location_col = "location",
  index = "date",
  key = "location"
)
}
\arguments{
\item{hub_data}{Data frame in hub format with columns for date, location, value, etc.}

\item{date_col}{Character. Name of the date column. Default is "date".}

\item{value_col}{Character. Name of the value column. Default is "value".}

\item{location_col}{Character. Name of the location column. Default is "location".}

\item{index}{Character. Name to use for the time index. Default is "date".}

\item{key}{Character vector. Key variables for tsibble. Default is "location".}
}
\value{
A tsibble object
}
\description{
Converts data from hub format to tsibble format for use with fable package.
}
\examples{
\dontrun{
hub_data <- data.frame(
  date = seq.Date(as.Date("2023-01-01"), by = "week", length.out = 10),
  location = "US",
  value = rnorm(10, 100, 10)
)
ts_data <- convert_hub_to_tsibble(hub_data)
}
}
