% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_model_file}
\alias{read_model_file}
\title{Read model file helper}
\usage{
read_model_file(path)
}
\arguments{
\item{path}{character(1) path to file}
}
\description{
helper that picks the right reader
supports .parquet, .parquet.gz, .csv, .csv.gz
}
\examples{
\dontrun{
read_model_file("path/to/file.parquet")
read_model_file("path/to/file.csv.gz")
}

}
