% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_save_model_output.R
\name{save_model_output}
\alias{save_model_output}
\title{Check and Save Model Output for Forecasting Hub}
\usage{
save_model_output(model_name = "AMPH-SARIMA", fc_output, reference_date)
}
\arguments{
\item{model_name}{Character. Name of the forecasting model (e.g., "AMPH-SARIMA").}

\item{fc_output}{Data frame. Forecast output to be saved.}

\item{forecast_date}{Character. Date of the forecast in "YYYY-MM-DD" format.}
}
\value{
NULL (invisibly)
}
\description{
This function helps save the output of a forecasting model into a structured
directory format suitable for submission to a forecasting hub. It creates
necessary directories if they do not exist and saves the forecast output
as a CSV file named with the forecast date and model name.
}
\examples{
\dontrun{
library(AMPHForecastSuite)
setup_file_structure(project_dir = "My_Forecast_Project")
}

}
