% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R, R/setup_file_structure.R
\name{setup_file_structure}
\alias{setup_file_structure}
\title{Set Up File Structure for a Forecasting Project}
\usage{
setup_file_structure(project_dir = "AMPH_forecast_project")

setup_file_structure(project_dir = "AMPH_forecast_project")
}
\arguments{
\item{project_dir}{Character. Directory to set up the project structure. Default is "AMPH_forecast_project".}
}
\value{
NULL (invisibly)

NULL (invisibly)
}
\description{
This function helps install all the necessary packages for building an
infectious disease forecasting pipeline and hub, including hubverse packages,
fable, EpiEstim, EpiNow2, epidatr, and epiforecast packages.

This function helps install all the necessary packages for building an
infectious disease forecasting pipeline and hub, including hubverse packages,
fable, EpiEstim, EpiNow2, epidatr, and epiforecast packages.
}
\examples{
\dontrun{
library(AMPHForecastSuite)
setup_file_structure(project_dir = "My_Forecast_Project")
}

\dontrun{
library(AMPHForecastSuite)
setup_file_structure(project_dir = "My_Forecast_Project")
}

}
