% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_epipredarx_hv.R
\name{trans_epipredarx_hv}
\alias{trans_epipredarx_hv}
\title{Transform Epipredict ARX Output to appropriate hubVerse format}
\usage{
trans_epipredarx_hv(
  fc_output = arx_forecast,
  model_name = "AMPH-epipredict-arx",
  target = "wk inc flu hosp",
  reference_date,
  horizon_time_steps = 0:3
)
}
\arguments{
\item{fc_output}{List. Output from the ARX forecasting model (e.g., arx_forecast).}

\item{model_name}{Character. Name of the forecasting model. Default is "AMPH-epipredict-arx".}

\item{target}{Character. Target variable name. Default is "wk inc flu hosp".}

\item{reference_date}{Date. The date of the forecast reference. This should be in "YYYY-MM-DD" format.}
}
\value{
tibble. A data frame in hub format.
}
\description{
Converts the output from an ARX forecasting model (e.g., arx_forecast)
to the required hub format for submission to a forecasting hub.
}
\examples{
\dontrun{
library(AMPHForecastSuite)
setup_file_structure(project_dir = "My_Forecast_Project")
# Assuming you have ARX forecast output
hub_forecast <- trans_epipredarx_hv(fc_output = arx_forecast,
                    model_name = "AMPH-epipredict-arx",
                    target = "wk inc flu hosp",
                    reference_date = forecast_date + 6,
                    horizon_time_steps = 0:3)
}

}
