% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_epipredclim_hv.R
\name{trans_epipredclim_hv}
\alias{trans_epipredclim_hv}
\title{Transform Epipredict Climate Output to appropriate hubVerse format}
\usage{
trans_epipredclim_hv(
  fc_output = climate_forecast,
  model_name = "AMPH-epipredict-climate",
  target = "wk inc flu hosp",
  reference_date,
  horizon_time_steps = 0:3
)
}
\arguments{
\item{fc_output}{List. The output from the climatological_forecaster model, typically containing a 'predictions' data frame.}

\item{model_name}{Character. Name of the forecasting model. Default is "AMPH-epipredict-climate".}

\item{target}{Character. Target variable name. Default is "wk inc flu hosp".}

\item{reference_date}{Date. The date of the forecast reference. This should be in "YYYY-MM-DD" format.}

\item{horizon_time_steps}{Integer vector. Time steps for the forecast horizon. Default is 0:3 (0 to 3 weeks ahead).}
}
\value{
tibble. A data frame in hub format.
}
\description{
Converts the output from a climatological_forecaster model
to the required hub format for submission to a forecasting hub.
}
\examples{
\dontrun{
library(AMPHForecastSuite)
setup_file_structure(project_dir = "My_Forecast_Project")
# Assuming you have climate_forecast output from climatological_forecaster hub_format
hub_forecast <- trans_epipredclim_hv(fc_output = climate_forecast,
                                    model_name = "AMPH-epipredict-climate",
                                    target = "wk inc flu hosp",
                                    reference_date = forecast_date + 6,
                                    horizon_time_steps = 0:3)
}

}
